use u16 for ArtnetEvent and add value ui field for changing universe and channels

+++ src/input/event.rs
@@ -7,7 +7,7 @@ use strum::Display;
 pub enum InputEvent {
     Key(Key),
     Gamepad(GamepadEvent),
-    Artnet(u8),
+    Artnet(u16),
 }
 
 impl InputEvent {
